Rewrite auto-role assignments to not use the cache
With the update to Discords API, bots who are not verified will no longer get Discord user event updates, which means the cache for users won't change once the first entry has been added to it, which then caused the bot to give the user the same roles over and over again, even giving roles that it shouldn't, with this change the bot should now only give roles to users who are missing the roles being checked for, while not using the cache at all.
With the change the role assignment job have been removed, so roles are now only given when a user verifies with the bot, or when they're selected by the Hypixel Rank Sync job, which runs every 30 second per user, so the delay might be really large with a lot of users.